@@ -3,7 +3,7 @@ name = "pest-language-server"
 version = "0.3.12"
 authors = ["Jamalam <james@jamalam.tech>"]
 description = "A language server for Pest."
-edition = "2021"
+edition = "2024"
 homepage = "https://pest.rs"
 license = "Apache-2.0"
 readme = "../README.md"
@@ -12,15 +12,15 @@ repository = "https://github.com/pest-parser/pest-ide-tools"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clap = { version = "4.5.48", features = ["cargo"] }
 pest = "2.8.2"
 pest_fmt = "0.2.5"
 pest_meta = {version = "2.8.2", features = ["grammar-extras"]}
-reqwest = { version = "0.12.4", features = [
-  "json",
-  "rustls-tls",
-], default-features = false }
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.116"
-tokio = { version = "1.37.0", features = ["full"] }
-tower-lsp = "0.20.0"
+smol = "2.0.2"
+strum = "0.27.2"
+strum_macros = "0.27.2"
+tower-lsp = { version = "0.20.0", features = ["runtime-agnostic"], default-features = false }
 unicode-segmentation = "1.11.0"
+url = { version = "2.5.7", features = ["serde"] }
