@@ -1,70 +1,73 @@
-use std::{collections::HashMap, str::Split};
+use std::{
+    collections::HashMap,
+    iter,
+    str::{FromStr, Split}
+};
 
-use pest_meta::parser;
+use pest_meta::parser::{self, Rule};
+use serde::Deserialize;
+use strum::IntoEnumIterator;
 use tower_lsp::{
-    jsonrpc::Result,
+    Client, jsonrpc,
     lsp_types::{
-        request::{GotoDeclarationParams, GotoDeclarationResponse},
         CodeAction, CodeActionKind, CodeActionOrCommand, CodeActionParams, CodeActionResponse,
         CompletionItem, CompletionItemKind, CompletionParams, CompletionResponse,
         ConfigurationItem, DeleteFilesParams, Diagnostic, DiagnosticSeverity,
         DidChangeConfigurationParams, DidChangeTextDocumentParams, DidChangeWatchedFilesParams,
-        DidOpenTextDocumentParams, DocumentChanges, DocumentFormattingParams, FileChangeType,
-        FileDelete, FileEvent, GotoDefinitionParams, GotoDefinitionResponse, Hover, HoverContents,
-        HoverParams, InitializedParams, Location, MarkedString, MessageType, OneOf,
+        DidOpenTextDocumentParams, DocumentChanges, DocumentFormattingParams, Documentation,
+        FileChangeType, FileEvent, Hover, HoverContents, HoverParams, InitializedParams, Location,
+        MarkedString, MarkupContent, MarkupKind, MessageType, OneOf,
         OptionalVersionedTextDocumentIdentifier, Position, PublishDiagnosticsParams, Range,
-        ReferenceParams, RenameParams, TextDocumentEdit, TextDocumentItem, TextEdit, Url,
-        VersionedTextDocumentIdentifier, WorkspaceEdit,
-    },
-    Client,
+        ReferenceParams, RenameParams, TextDocumentEdit, TextDocumentIdentifier, TextDocumentItem,
+        TextDocumentPositionParams, TextEdit, Url, VersionedTextDocumentIdentifier, WorkspaceEdit
+    }
 };
 use unicode_segmentation::UnicodeSegmentation;
 
 use crate::{
     analysis::{Analysis, RuleAnalysis},
-    builtins::BUILTINS,
-    config::Config,
+    builtins::Builtin,
     helpers::{
-        create_empty_diagnostics, range_contains, str_range, validate_pairs, Diagnostics,
-        Documents, FindWordRange, IntoDiagnostics, IntoRangeWithLine,
-    },
+        Diagnostics, Documents, FindWordRange, IntoDiagnostics, IntoRangeWithLine, RangeContains,
+        create_empty_diagnostics, str_range, validate_pairs
+    }
 };
-use crate::{builtins::get_builtin_description, update_checker::check_for_updates};
+
+#[derive(Deserialize, Default, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct Config {
+    pub always_used_rule_names: Vec<String>
+}
 
 #[derive(Debug)]
 pub struct PestLanguageServerImpl {
     pub client: Client,
     pub documents: Documents,
     pub analyses: HashMap<Url, Analysis>,
-    pub config: Config,
+    pub config: Config
 }
 
 impl PestLanguageServerImpl {
-    pub async fn initialized(&mut self, _: InitializedParams) {
-        let config_items = self
+    async fn try_update_config(&mut self) -> Option<Config> {
+        let value = self
             .client
             .configuration(vec![ConfigurationItem {
                 scope_uri: None,
-                section: Some("pestIdeTools".to_string()),
+                section: Some("pestIdeTools".to_string())
             }])
-            .await;
+            .await
+            .ok()?
+            .into_iter()
+            .next()?;
+        serde_json::from_value(value).ok()
+    }
 
-        let mut updated_config = false;
-
-        if let Ok(config_items) = config_items {
-            if let Some(config) = config_items.into_iter().next() {
-                if let Ok(config) = serde_json::from_value(config) {
-                    self.config = config;
-                    updated_config = true;
-                }
-            }
-        }
-
-        if !updated_config {
+    pub async fn initialized(&mut self, _: InitializedParams) {
+        if self.try_update_config().await.is_none() {
             self.client
                 .log_message(
                     MessageType::ERROR,
-                    "Failed to retrieve configuration from client.",
+                    "Failed to retrieve configuration from client"
                 )
                 .await;
         }
@@ -72,30 +75,12 @@ impl PestLanguageServerImpl {
         self.client
             .log_message(
                 MessageType::INFO,
-                format!("Pest Language Server v{}", env!("CARGO_PKG_VERSION")),
+                format!("Pest Language Server v{}", env!("CARGO_PKG_VERSION"))
             )
             .await;
-
-        if self.config.check_for_updates {
-            self.client
-                .log_message(MessageType::INFO, "Checking for updates...".to_string())
-                .await;
-
-            if let Some(new_version) = check_for_updates().await {
-                self.client
-                    .show_message(
-                        MessageType::INFO,
-                        format!(
-                            "A new version of the Pest Language Server is available: v{}",
-                            new_version
-                        ),
-                    )
-                    .await;
-            }
-        }
     }
 
-    pub async fn shutdown(&self) -> Result<()> {
+    pub async fn shutdown(&self) -> jsonrpc::Result<()> {
         self.client
             .log_message(MessageType::INFO, "Pest Language Server shutting down :)")
             .await;
@@ -103,18 +88,20 @@ impl PestLanguageServerImpl {
     }
 
     pub async fn did_change_configuration(&mut self, params: DidChangeConfigurationParams) {
-        if let Ok(config) = serde_json::from_value(params.settings) {
-            self.config = config;
-            self.client
-                .log_message(
-                    MessageType::INFO,
-                    "Updated configuration from client.".to_string(),
-                )
-                .await;
+        let Ok(config) = serde_json::from_value(params.settings) else {
+            return;
+        };
 
-            let diagnostics = self.reload().await;
-            self.send_diagnostics(diagnostics).await;
-        }
+        self.config = config;
+        self.client
+            .log_message(
+                MessageType::INFO,
+                "Updated configuration from client.".to_string()
+            )
+            .await;
+
+        let diagnostics = self.reload().await;
+        self.send_diagnostics(diagnostics).await;
     }
 
     pub async fn did_open(&mut self, params: DidOpenTextDocumentParams) {
@@ -127,7 +114,7 @@ impl PestLanguageServerImpl {
             self.client
                 .log_message(
                     MessageType::INFO,
-                    "\tReopened already tracked document.".to_string(),
+                    "\tReopened already tracked document.".to_string()
                 )
                 .await;
         }
@@ -139,28 +126,26 @@ impl PestLanguageServerImpl {
     pub async fn did_change(&mut self, params: DidChangeTextDocumentParams) {
         let DidChangeTextDocumentParams {
             text_document,
-            content_changes,
+            content_changes
         } = params;
         let VersionedTextDocumentIdentifier { uri, version } = text_document;
 
-        assert_eq!(content_changes.len(), 1);
-        let change = content_changes.into_iter().next().unwrap();
-        assert!(change.range.is_none());
+        for change in content_changes.into_iter() {
+            let updated_doc =
+                TextDocumentItem::new(uri.clone(), "pest".to_owned(), version, change.text);
 
-        let updated_doc =
-            TextDocumentItem::new(uri.clone(), "pest".to_owned(), version, change.text);
+            if self.upsert_document(updated_doc).is_none() {
+                self.client
+                    .log_message(
+                        MessageType::INFO,
+                        format!("Updated untracked document {}", uri)
+                    )
+                    .await;
+            }
 
-        if self.upsert_document(updated_doc).is_none() {
-            self.client
-                .log_message(
-                    MessageType::INFO,
-                    format!("Updated untracked document {}", uri),
-                )
-                .await;
+            let diagnostics = self.reload().await;
+            self.send_diagnostics(diagnostics).await;
         }
-
-        let diagnostics = self.reload().await;
-        self.send_diagnostics(diagnostics).await;
     }
 
     pub async fn did_change_watched_files(&mut self, params: DidChangeWatchedFilesParams) {
@@ -179,27 +164,27 @@ impl PestLanguageServerImpl {
             self.client
                 .log_message(
                     MessageType::INFO,
-                    format!("Deleting removed document {}", uri),
+                    format!("Deleting removed document {}", uri)
                 )
                 .await;
 
-            if let Some(removed) = self.remove_document(&uri) {
-                let (_, empty_diagnostics) = create_empty_diagnostics((&uri, &removed));
-                if diagnostics.insert(uri, empty_diagnostics).is_some() {
+            match self.remove_document(&uri) {
+                Some(removed) => {
+                    let empty_diagnostics = create_empty_diagnostics(uri.clone(), &removed);
+                    if diagnostics.insert(uri, empty_diagnostics).is_some() {
+                        self.client
+                            .log_message(MessageType::WARNING, "\tDuplicate URIs in event payload")
+                            .await;
+                    }
+                }
+                None => {
                     self.client
                         .log_message(
                             MessageType::WARNING,
-                            "\tDuplicate URIs in event payload".to_string(),
+                            "\tAttempted to delete untracked document"
                         )
                         .await;
                 }
-            } else {
-                self.client
-                    .log_message(
-                        MessageType::WARNING,
-                        "\tAttempted to delete untracked document".to_string(),
-                    )
-                    .await;
             }
         }
 
@@ -208,14 +193,14 @@ impl PestLanguageServerImpl {
     }
 
     pub async fn did_delete_files(&mut self, params: DeleteFilesParams) {
-        let DeleteFilesParams { files } = params;
-        let mut uris = vec![];
-        for FileDelete { uri } in files {
-            match Url::parse(&uri) {
+        let files = params.files;
+        let mut uris = Vec::with_capacity(files.len());
+        for file in files {
+            match Url::parse(&file.uri) {
                 Ok(uri) => uris.push(uri),
                 Err(e) => {
                     self.client
-                        .log_message(MessageType::ERROR, format!("Failed to parse URI {}", e))
+                        .log_message(MessageType::ERROR, format!("Failed to parse URI {e}"))
                         .await
                 }
             }
@@ -229,20 +214,18 @@ impl PestLanguageServerImpl {
 
         for uri in uris {
             let removed = self.remove_documents_in_dir(&uri);
-            if !removed.is_empty() {
-                for (uri, params) in removed {
-                    self.client
-                        .log_message(MessageType::INFO, format!("\tDeleted {}", uri))
-                        .await;
+            for (uri, params) in removed {
+                self.client
+                    .log_message(MessageType::INFO, format!("\tDeleted {uri}"))
+                    .await;
 
-                    if diagnostics.insert(uri, params).is_some() {
-                        self.client
-                            .log_message(
-                                MessageType::INFO,
-                                "\tDuplicate URIs in event payload".to_string(),
-                            )
-                            .await;
-                    }
+                if diagnostics.insert(uri, params).is_some() {
+                    self.client
+                        .log_message(
+                            MessageType::INFO,
+                            "\tDuplicate URIs in event payload".to_string()
+                        )
+                        .await;
                 }
             }
         }
@@ -253,255 +236,207 @@ impl PestLanguageServerImpl {
         }
     }
 
-    pub fn code_action(&self, params: CodeActionParams) -> Result<Option<CodeActionResponse>> {
+    pub fn code_action(&self, params: CodeActionParams) -> Option<CodeActionResponse> {
         let CodeActionParams {
             context,
             range,
             text_document,
             ..
         } = params;
-        let only = context.only;
-        let analysis = self.analyses.get(&text_document.uri);
-        let mut actions = Vec::new();
+        let only = context.only.as_ref();
+        let analysis = self.analyses.get(&text_document.uri)?;
+        let inline_all = only
+            .is_none_or(|only| only.contains(&CodeActionKind::REFACTOR_INLINE))
+            .then(|| self.refactor_inline_all(analysis, &text_document, range))
+            .flatten();
 
-        if let Some(analysis) = analysis {
-            // Inlining
-            if only
-                .as_ref()
-                .is_none_or(|only| only.contains(&CodeActionKind::REFACTOR_INLINE))
-            {
-                let mut rule_name = None;
+        let inline = only
+            .is_none_or(|only| only.contains(&CodeActionKind::REFACTOR_INLINE))
+            .then(|| self.refactor_inline(analysis, &text_document, range))
+            .flatten();
 
-                for (name, ra) in analysis.rules.iter() {
-                    if let Some(ra) = ra {
-                        if ra.identifier_location.range == range {
-                            rule_name = Some(name);
-                            break;
-                        }
-                    }
-                }
+        let extract = only
+            .is_none_or(|only| only.contains(&CodeActionKind::REFACTOR_EXTRACT))
+            .then(|| self.refactor_extract(range, text_document))
+            .flatten();
 
-                if let Some(rule_name) = rule_name {
-                    let ra = self
-                        .get_rule_analysis(&text_document.uri, rule_name)
-                        .expect("should not be called on a builtin with no rule analysis");
-                    let rule_expression = ra.expression.clone();
-
-                    let mut edits = Vec::new();
-
-                    edits.push(TextEdit {
-                        range: ra.definition_location.range,
-                        new_text: "".to_owned(),
-                    });
-
-                    if let Some(occurrences) = self
-                        .get_rule_analysis(&text_document.uri, rule_name)
-                        .map(|ra| &ra.occurrences)
-                    {
-                        for occurrence in occurrences {
-                            if occurrence.range != ra.identifier_location.range {
-                                edits.push(TextEdit {
-                                    range: occurrence.range,
-                                    new_text: rule_expression.clone(),
-                                });
-                            }
-                        }
-                    }
-
-                    let mut changes = HashMap::new();
-                    changes.insert(text_document.uri.clone(), edits);
-
-                    actions.push(CodeActionOrCommand::CodeAction(CodeAction {
-                        title: "Inline rule".to_owned(),
-                        kind: Some(CodeActionKind::REFACTOR_INLINE),
-                        edit: Some(WorkspaceEdit {
-                            changes: Some(changes),
-                            document_changes: None,
-                            change_annotations: None,
-                        }),
-                        ..Default::default()
-                    }));
-                }
-            }
-
-            if only
-                .as_ref()
-                .is_none_or(|only| only.contains(&CodeActionKind::REFACTOR_EXTRACT))
-                && range.start.line == range.end.line
-            {
-                let document = self.documents.get(&text_document.uri).unwrap();
-                let mut lines = document.text.lines();
-                let line = lines.nth(range.start.line as usize).unwrap_or("");
-
-                let mut rule_name_start_idx = 0;
-                let mut chars = line.graphemes(true);
-
-                while chars.next() == Some(" ") {
-                    rule_name_start_idx += 1;
-                }
-
-                let name_range = line.get_word_range_at_idx(rule_name_start_idx);
-                let rule_name = &str_range(line, &name_range);
-
-                if let Some(ra) = self.get_rule_analysis(&text_document.uri, rule_name) {
-                    let mut selected_token = None;
-
-                    for (token, location) in ra.tokens.iter() {
-                        if range_contains(&location.range, &range) {
-                            selected_token = Some((token, location));
-                            break;
-                        }
-                    }
-
-                    if let Some((extracted_token, location)) = selected_token {
-                        //TODO: Replace with something more robust, it's horrible
-                        let extracted_token_identifier =
-                            match parser::parse(parser::Rule::node, extracted_token) {
-                                Ok(mut node) => {
-                                    let mut next = node.next().unwrap();
-
-                                    loop {
-                                        match next.as_rule() {
-                                            parser::Rule::terminal => {
-                                                next = next.into_inner().next().unwrap();
-                                            }
-                                            parser::Rule::_push
-                                            | parser::Rule::peek_slice
-                                            | parser::Rule::identifier
-                                            | parser::Rule::string
-                                            | parser::Rule::insensitive_string
-                                            | parser::Rule::range => {
-                                                break Some(next);
-                                            }
-                                            parser::Rule::opening_paren => {
-                                                node = node.next().unwrap().into_inner();
-                                                let next_opt = node
-                                                    .find(|r| r.as_rule() == parser::Rule::term)
-                                                    .unwrap()
-                                                    .into_inner()
-                                                    .find(|r| r.as_rule() == parser::Rule::node);
-
-                                                if let Some(new_next) = next_opt {
-                                                    next = new_next;
-                                                } else {
-                                                    break None;
-                                                }
-                                            }
-                                            _ => unreachable!(
-                                                "unexpected rule in node: {:?}",
-                                                next.as_rule()
-                                            ),
-                                        };
-                                    }
-                                    .map(|p| p.as_str())
-                                }
-                                Err(_) => None,
-                            }
-                            .unwrap_or("");
-
-                        if self
-                            .get_rule_analysis(&text_document.uri, extracted_token_identifier)
-                            .is_some()
-                            || BUILTINS.contains(&extracted_token_identifier)
-                            || extracted_token.starts_with('\"')
-                            || extracted_token.starts_with('\'')
-                            || extracted_token.starts_with("PUSH")
-                            || extracted_token.starts_with("PEEK")
-                            || extracted_token.starts_with("^\"")
-                        {
-                            let mut rule_name_number = 0;
-                            let extracted_rule_name = loop {
-                                rule_name_number += 1;
-                                let extracted_rule_name =
-                                    format!("{}_{}", rule_name, rule_name_number);
-                                if self
-                                    .get_rule_analysis(&text_document.uri, &extracted_rule_name)
-                                    .is_none()
-                                {
-                                    break extracted_rule_name;
-                                }
-                            };
-
-                            let extracted_rule = format!(
-                                "{} = {{ {} }}",
-                                extracted_rule_name.trim(),
-                                extracted_token.trim(),
-                            );
-
-                            let mut edits = Vec::new();
-
-                            edits.push(TextEdit {
-                                range: Range {
-                                    start: Position {
-                                        line: location.range.end.line + 1,
-                                        character: 0,
-                                    },
-                                    end: Position {
-                                        line: location.range.end.line + 1,
-                                        character: 0,
-                                    },
-                                },
-                                new_text: format!(
-                                    "{}{}\n",
-                                    if line.ends_with('\n') { "" } else { "\n" },
-                                    extracted_rule
-                                ),
-                            });
-
-                            let mut changes = HashMap::new();
-                            changes.insert(text_document.uri.clone(), edits);
-
-                            for (url, analysis) in self.analyses.iter() {
-                                for (_, ra) in analysis.rules.iter() {
-                                    if let Some(ra) = ra {
-                                        for (token, location) in ra.tokens.iter() {
-                                            if token == extracted_token {
-                                                changes
-                                                    .entry(url.clone())
-                                                    .or_insert_with(Vec::new)
-                                                    .push(TextEdit {
-                                                        range: location.range,
-                                                        new_text: format!(
-                                                            "{} ",
-                                                            extracted_rule_name
-                                                        ),
-                                                    });
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-
-                            actions.push(CodeActionOrCommand::CodeAction(CodeAction {
-                                title: "Extract into new rule".to_owned(),
-                                kind: Some(CodeActionKind::REFACTOR_EXTRACT),
-                                edit: Some(WorkspaceEdit {
-                                    changes: Some(changes),
-                                    document_changes: None,
-                                    change_annotations: None,
-                                }),
-                                ..Default::default()
-                            }));
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(Some(actions))
+        let actions = inline_all
+            .into_iter()
+            .chain(extract)
+            .chain(inline)
+            .map(CodeActionOrCommand::CodeAction)
+            .collect();
+        Some(actions)
     }
 
-    pub fn completion(&self, params: CompletionParams) -> Result<Option<CompletionResponse>> {
+    fn refactor_inline(
+        &self,
+        analysis: &Analysis,
+        text_document: &TextDocumentIdentifier,
+        range: Range
+    ) -> Option<CodeAction> {
+        let ((name, ra), occurence) = analysis.rules.iter().find_map(|ra| {
+            Some((
+                ra,
+                ra.1.occurrences
+                    .iter()
+                    .filter(|occurence| *occurence != &ra.1.identifier_location)
+                    .find(|occurence| occurence.range.contains(range))?
+            ))
+        })?;
+
+        let edit = vec![TextEdit {
+            range: occurence.range,
+            new_text: ra.expression.clone()
+        }];
+
+        let change = HashMap::from_iter(iter::once((text_document.uri.clone(), edit)));
+
+        let edit = Some(WorkspaceEdit {
+            changes: Some(change),
+            document_changes: None,
+            change_annotations: None
+        });
+
+        Some(CodeAction {
+            title: format!("Inline {name}"),
+            kind: Some(CodeActionKind::REFACTOR_INLINE),
+            edit,
+            ..Default::default()
+        })
+    }
+
+    fn refactor_inline_all(
+        &self,
+        analysis: &Analysis,
+        text_document: &TextDocumentIdentifier,
+        range: Range
+    ) -> Option<CodeAction> {
+        let (name, ra) = analysis.rules.iter().find(|ra| {
+            ra.1.occurrences
+                .iter()
+                .any(|occurence| occurence.range.contains(range))
+        })?;
+
+        if ra.occurrences.len() == 1 {
+            return None;
+        }
+
+        let rule_expression = ra.expression.clone();
+
+        let edits = ra
+            .occurrences
+            .iter()
+            .filter(|occurence| ra.identifier_location.range != occurence.range)
+            .map(|occurrence| TextEdit {
+                range: occurrence.range,
+                new_text: rule_expression.clone()
+            })
+            .chain(iter::once(TextEdit {
+                range: ra.definition_location.range,
+                new_text: String::new()
+            }))
+            .collect();
+
+        let changes = HashMap::from_iter(iter::once((text_document.uri.clone(), edits)));
+
+        let edit = Some(WorkspaceEdit {
+            changes: Some(changes),
+            document_changes: None,
+            change_annotations: None
+        });
+
+        Some(CodeAction {
+            title: format!("Inline all occurrences of {name}"),
+            kind: Some(CodeActionKind::REFACTOR_INLINE),
+            edit,
+            ..Default::default()
+        })
+    }
+
+    fn refactor_extract(
+        &self,
+        range: Range,
+        text_document: TextDocumentIdentifier
+    ) -> Option<CodeAction> {
+        let document = &self.documents[&text_document.uri];
+        let line = document
+            .text
+            .lines()
+            .nth(range.start.line as usize)
+            .unwrap_or("");
+
+        let rule_name_start_idx = line
+            .graphemes(true)
+            .take_while(|grapheme| *grapheme == " ")
+            .count();
+
+        let name_range = line.get_word_range_at_idx(rule_name_start_idx);
+        let rule_name = &str_range(line, &name_range);
+
+        let ra = self.get_rule_analysis(&text_document.uri, rule_name)?;
+        let selected_token = ra
+            .tokens
+            .iter()
+            .find(|(_, location)| location.range.contains(range));
+
+        let (extracted_token, extracted_token_location) = selected_token?;
+        let extracted_rule_name = (0..)
+            .map(|rule_name_number| format!("{}_{}", rule_name, rule_name_number))
+            .find(|extracted_rule_name| {
+                self.get_rule_analysis(&text_document.uri, extracted_rule_name)
+                    .is_none()
+            })?;
+
+        let extracted_rule = format!(
+            "{} = {{ {} }}",
+            extracted_rule_name.trim(),
+            extracted_token.trim(),
+        );
+
+        let pos = Position {
+            line: extracted_token_location.range.end.line + 1,
+            character: 0
+        };
+
+        let prefix = if line.ends_with('\n') { "" } else { "\n" };
+        let edits = vec![
+            TextEdit {
+                range: Range {
+                    start: pos,
+                    end: pos
+                },
+                new_text: format!("{prefix}{extracted_rule}\n",)
+            },
+            TextEdit {
+                range: extracted_token_location.range,
+                new_text: extracted_rule_name.clone()
+            },
+        ];
+
+        let changes = HashMap::from_iter(iter::once((text_document.uri.clone(), edits)));
+
+        let edit = Some(WorkspaceEdit {
+            changes: Some(changes),
+            document_changes: None,
+            change_annotations: None
+        });
+
+        Some(CodeAction {
+            title: format!("Extract {rule_name} into new rule ({extracted_rule_name})"),
+            kind: Some(CodeActionKind::REFACTOR_EXTRACT),
+            edit,
+            ..Default::default()
+        })
+    }
+
+    pub fn completion(&self, params: CompletionParams) -> Option<CompletionResponse> {
         let CompletionParams {
             text_document_position,
             ..
         } = params;
 
-        let document = self
-            .documents
-            .get(&text_document_position.text_document.uri)
-            .unwrap();
-
+        let document = &self.documents[&text_document_position.text_document.uri];
         let mut lines = document.text.lines();
         let line = lines
             .nth(text_document_position.position.line as usize)
@@ -509,33 +444,42 @@ impl PestLanguageServerImpl {
         let range = line.get_word_range_at_idx(text_document_position.position.character as usize);
         let partial_identifier = &str_range(line, &range);
 
-        if let Some(analysis) = self.analyses.get(&document.uri) {
-            return Ok(Some(CompletionResponse::Array(
-                analysis
-                    .rules
-                    .keys()
-                    .filter(|i| partial_identifier.is_empty() || i.starts_with(partial_identifier))
-                    .map(|i| CompletionItem {
-                        label: i.to_owned(),
-                        kind: Some(CompletionItemKind::FIELD),
-                        ..Default::default()
+        let analysis = self.analyses.get(&document.uri)?;
+        let rule_completions = analysis
+            .rules
+            .iter()
+            .filter(|(i, _)| i.starts_with(partial_identifier))
+            .map(|(i, ra)| CompletionItem {
+                label: i.to_owned(),
+                kind: Some(CompletionItemKind::FIELD),
+                documentation: ra
+                    .doc
+                    .clone()
+                    .map(|value| MarkupContent {
+                        kind: MarkupKind::Markdown,
+                        value
                     })
-                    .collect(),
-            )));
-        }
+                    .map(Documentation::MarkupContent),
+                ..Default::default()
+            });
 
-        Ok(None)
+        let builtins_completions = Builtin::iter().map(|builtin| CompletionItem {
+            label: builtin.as_ref().to_string(),
+            kind: Some(builtin.kind()),
+            documentation: Some(Documentation::String(builtin.description().to_string())),
+            ..Default::default()
+        });
+
+        let completions = rule_completions.chain(builtins_completions).collect();
+        Some(CompletionResponse::Array(completions))
     }
 
-    pub fn hover(&self, params: HoverParams) -> Result<Option<Hover>> {
+    pub fn hover(&self, params: HoverParams) -> Option<Hover> {
         let HoverParams {
             text_document_position_params,
             ..
         } = params;
-        let document = self
-            .documents
-            .get(&text_document_position_params.text_document.uri)
-            .unwrap();
+        let document = &self.documents[&text_document_position_params.text_document.uri];
 
         let mut lines = document.text.lines();
         let line = lines
@@ -545,38 +489,29 @@ impl PestLanguageServerImpl {
             line.get_word_range_at_idx(text_document_position_params.position.character as usize);
         let identifier = &str_range(line, &range);
 
-        if let Some(description) = get_builtin_description(identifier) {
-            return Ok(Some(Hover {
-                contents: HoverContents::Scalar(MarkedString::String(description.to_owned())),
-                range: Some(range.into_range(text_document_position_params.position.line)),
-            }));
+        if let Ok(builtin) = Builtin::from_str(identifier) {
+            let contents =
+                HoverContents::Scalar(MarkedString::String(builtin.description().to_owned()));
+            let range = Some(range.into_range(text_document_position_params.position.line));
+            let hover = Hover { contents, range };
+            return Some(hover);
         }
 
-        if let Some(Some(ra)) = self
-            .analyses
-            .get(&document.uri)
-            .and_then(|a| a.rules.get(identifier))
-        {
-            return Ok(Some(Hover {
-                contents: HoverContents::Scalar(MarkedString::String(ra.doc.clone())),
-                range: Some(range.into_range(text_document_position_params.position.line)),
-            }));
-        }
+        let ra = self.analyses.get(&document.uri)?.rules.get(identifier)?;
 
-        Ok(None)
+        let contents = HoverContents::Scalar(MarkedString::String(ra.doc.clone()?));
+        let range = Some(range.into_range(text_document_position_params.position.line));
+        Some(Hover { contents, range })
     }
 
-    pub fn rename(&self, params: RenameParams) -> Result<Option<WorkspaceEdit>> {
+    pub fn rename(&self, params: RenameParams) -> WorkspaceEdit {
         let RenameParams {
             text_document_position,
             new_name,
             ..
         } = params;
 
-        let document = self
-            .documents
-            .get(&text_document_position.text_document.uri)
-            .unwrap();
+        let document = &self.documents[&text_document_position.text_document.uri];
         let line = document
             .text
             .lines()
@@ -584,116 +519,60 @@ impl PestLanguageServerImpl {
             .unwrap_or("");
         let old_identifier = &str_range(
             line,
-            &line.get_word_range_at_idx(text_document_position.position.character as usize),
+            &line.get_word_range_at_idx(text_document_position.position.character as usize)
         );
-        let mut edits = Vec::new();
 
-        if let Some(occurrences) = self
+        let edits = self
             .get_rule_analysis(&document.uri, old_identifier)
-            .map(|ra| &ra.occurrences)
-        {
-            for location in occurrences {
-                edits.push(TextEdit {
-                    range: location.range,
-                    new_text: new_name.clone(),
-                });
-            }
-        }
+            .into_iter()
+            .flat_map(|ra| &ra.occurrences)
+            .map(|location| TextEdit {
+                range: location.range,
+                new_text: new_name.clone()
+            })
+            .map(OneOf::Left)
+            .collect();
 
-        Ok(Some(WorkspaceEdit {
+        let text_document = OptionalVersionedTextDocumentIdentifier {
+            uri: text_document_position.text_document.uri,
+            version: Some(document.version)
+        };
+
+        let edit = TextDocumentEdit {
+            text_document,
+            edits
+        };
+
+        let document_changes = Some(DocumentChanges::Edits(vec![edit]));
+
+        WorkspaceEdit {
             change_annotations: None,
             changes: None,
-            document_changes: Some(DocumentChanges::Edits(vec![TextDocumentEdit {
-                text_document: OptionalVersionedTextDocumentIdentifier {
-                    uri: text_document_position.text_document.uri,
-                    version: Some(document.version),
-                },
-                edits: edits.into_iter().map(OneOf::Left).collect(),
-            }])),
-        }))
+            document_changes
+        }
     }
 
-    pub fn goto_declaration(
-        &self,
-        params: GotoDeclarationParams,
-    ) -> Result<Option<GotoDeclarationResponse>> {
-        let GotoDeclarationParams {
-            text_document_position_params,
-            ..
-        } = params;
-
-        let document = self
-            .documents
-            .get(&text_document_position_params.text_document.uri)
-            .unwrap();
-
+    pub fn goto_definition(&self, params: TextDocumentPositionParams) -> Option<Location> {
+        let uri = params.text_document.uri;
+        let document = &self.documents[&uri];
         let mut lines = document.text.lines();
-        let line = lines
-            .nth(text_document_position_params.position.line as usize)
-            .unwrap_or("");
-
-        let range =
-            line.get_word_range_at_idx(text_document_position_params.position.character as usize);
+        let line = lines.nth(params.position.line as usize).unwrap_or("");
+        let range = line.get_word_range_at_idx(params.position.character as usize);
         let identifier = &str_range(line, &range);
 
-        if let Some(location) = self
-            .get_rule_analysis(&document.uri, identifier)
-            .map(|ra| &ra.definition_location)
-        {
-            return Ok(Some(GotoDeclarationResponse::Scalar(Location {
-                uri: text_document_position_params.text_document.uri,
-                range: location.range,
-            })));
-        }
-
-        Ok(None)
+        let Location { range, .. } = self
+            .get_rule_analysis(&document.uri, identifier)?
+            .definition_location;
+        Some(Location { uri, range })
     }
 
-    pub fn goto_definition(
-        &self,
-        params: GotoDefinitionParams,
-    ) -> Result<Option<GotoDefinitionResponse>> {
-        let GotoDefinitionParams {
-            text_document_position_params,
-            ..
-        } = params;
-
-        let document = self
-            .documents
-            .get(&text_document_position_params.text_document.uri)
-            .unwrap();
-
-        let mut lines = document.text.lines();
-        let line = lines
-            .nth(text_document_position_params.position.line as usize)
-            .unwrap_or("");
-        let range =
-            line.get_word_range_at_idx(text_document_position_params.position.character as usize);
-        let identifier = &str_range(line, &range);
-
-        if let Some(location) = self
-            .get_rule_analysis(&document.uri, identifier)
-            .map(|ra| &ra.definition_location)
-        {
-            return Ok(Some(GotoDeclarationResponse::Scalar(Location {
-                uri: text_document_position_params.text_document.uri,
-                range: location.range,
-            })));
-        }
-
-        Ok(None)
-    }
-
-    pub fn references(&self, params: ReferenceParams) -> Result<Option<Vec<Location>>> {
+    pub fn references(&self, params: ReferenceParams) -> Option<Vec<Location>> {
         let ReferenceParams {
             text_document_position,
             ..
         } = params;
 
-        let document = self
-            .documents
-            .get(&text_document_position.text_document.uri)
-            .unwrap();
+        let document = &self.documents[&text_document_position.text_document.uri];
 
         let mut lines = document.text.lines();
         let line = lines
@@ -702,33 +581,67 @@ impl PestLanguageServerImpl {
         let range = line.get_word_range_at_idx(text_document_position.position.character as usize);
         let identifier = &str_range(line, &range);
 
-        Ok(self
-            .get_rule_analysis(&document.uri, identifier)
-            .map(|ra| ra.occurrences.clone()))
+        let rule_analysis = self.get_rule_analysis(&document.uri, identifier)?;
+        let locations = rule_analysis.occurrences.clone();
+        Some(locations)
     }
 
-    pub fn formatting(&self, params: DocumentFormattingParams) -> Result<Option<Vec<TextEdit>>> {
+    pub fn formatting(&self, params: DocumentFormattingParams) -> Option<Vec<TextEdit>> {
         let DocumentFormattingParams { text_document, .. } = params;
 
-        let document = self.documents.get(&text_document.uri).unwrap();
+        let document = &self.documents[&text_document.uri];
         let input = document.text.as_str();
 
         let fmt = pest_fmt::Formatter::new(input);
-        if let Ok(formatted) = fmt.format() {
-            let lines = document.text.lines();
-            let last_line = lines.clone().last().unwrap_or("");
-            let range = Range::new(
-                Position::new(0, 0),
-                Position::new(lines.count() as u32, last_line.len() as u32),
-            );
-            return Ok(Some(vec![TextEdit::new(range, formatted)]));
-        }
-
-        Ok(None)
+        let formatted = fmt.format().ok()?;
+        let lines = document.text.lines();
+        let last_line = lines.clone().last().unwrap_or("");
+        let end = Position::new(lines.count() as u32, last_line.len() as u32);
+        let range = Range::new(Position::new(0, 0), end);
+        Some(vec![TextEdit::new(range, formatted)])
     }
 }
 
 impl PestLanguageServerImpl {
+    async fn analyse_document(
+        analyses: &mut HashMap<Url, Analysis>,
+        diagnostics: &mut Diagnostics,
+        config: &Config,
+        url: Url,
+        document: &TextDocumentItem
+    ) -> Result<(), Vec<pest::error::Error<Rule>>> {
+        let pairs = parser::parse(parser::Rule::grammar_rules, document.text.as_str())
+            .map_err(|err| vec![err])?;
+        let analysis = analyses.entry(url.clone()).or_insert(Analysis {
+            doc_url: url.clone(),
+            rules: HashMap::new()
+        });
+        analysis.update_from(pairs.clone());
+
+        let unused_diagnostics: Vec<_> = analysis
+            .get_unused_rules()
+            .iter()
+            .filter(|(rule_name, _)| !config.always_used_rule_names.contains(rule_name))
+            .map(|(rule_name, rule_location)| Diagnostic {
+                range: rule_location.range,
+                severity: Some(DiagnosticSeverity::WARNING),
+                source: Some("Pest Language Server".to_owned()),
+                message: format!("Rule {} is unused", rule_name),
+                ..Default::default()
+            })
+            .collect();
+
+        if unused_diagnostics.len() > 1 {
+            diagnostics
+                .entry(url.clone())
+                .or_insert_with(|| create_empty_diagnostics(url, document))
+                .diagnostics
+                .extend(unused_diagnostics);
+        }
+
+        validate_pairs(pairs)
+    }
+
     async fn reload(&mut self) -> Diagnostics {
         self.client
             .log_message(MessageType::INFO, "Reloading all diagnostics".to_string())
@@ -739,74 +652,30 @@ impl PestLanguageServerImpl {
             self.client
                 .log_message(
                     MessageType::INFO,
-                    format!("\tReloading diagnostics for {}", url),
+                    format!("\tReloading diagnostics for {}", url)
                 )
                 .await;
 
-            let pairs = match parser::parse(parser::Rule::grammar_rules, document.text.as_str()) {
-                Ok(pairs) => Ok(pairs),
-                Err(error) => Err(vec![error]),
-            };
-
-            if let Ok(pairs) = pairs {
-                if let Err(errors) = validate_pairs(pairs.clone()) {
-                    diagnostics.insert(
-                        url.clone(),
-                        PublishDiagnosticsParams::new(
-                            url.clone(),
-                            errors.into_diagnostics(),
-                            Some(document.version),
-                        ),
-                    );
-                } else {
-                    let (_, empty_diagnostics) = create_empty_diagnostics((url, document));
-                    diagnostics.insert(url.clone(), empty_diagnostics);
-                }
-
-                self.analyses
-                    .entry(url.clone())
-                    .or_insert_with(|| Analysis {
-                        doc_url: url.clone(),
-                        rules: HashMap::new(),
-                    })
-                    .update_from(pairs);
-            } else if let Err(errors) = pairs {
-                diagnostics.insert(
+            if let Err(error) = Self::analyse_document(
+                &mut self.analyses,
+                &mut diagnostics,
+                &self.config,
+                url.clone(),
+                document
+            )
+            .await
+            {
+                let v = PublishDiagnosticsParams::new(
                     url.clone(),
-                    PublishDiagnosticsParams::new(
-                        url.clone(),
-                        errors.into_diagnostics(),
-                        Some(document.version),
-                    ),
+                    error.into_diagnostics(),
+                    Some(document.version)
                 );
-            }
 
-            if let Some(analysis) = self.analyses.get(url) {
-                let mut unused_diagnostics = Vec::new();
-                for (rule_name, rule_location) in
-                    analysis.get_unused_rules().iter().filter(|(rule_name, _)| {
-                        !self.config.always_used_rule_names.contains(rule_name)
-                    })
-                {
-                    unused_diagnostics.push(Diagnostic::new(
-                        rule_location.range,
-                        Some(DiagnosticSeverity::WARNING),
-                        None,
-                        Some("Pest Language Server".to_owned()),
-                        format!("Rule {} is unused", rule_name),
-                        None,
-                        None,
-                    ));
-                }
-
-                if unused_diagnostics.len() > 1 {
-                    diagnostics
-                        .entry(url.to_owned())
-                        .or_insert_with(|| create_empty_diagnostics((url, document)).1)
-                        .diagnostics
-                        .extend(unused_diagnostics);
-                }
-            }
+                diagnostics.insert(url.clone(), v)
+            } else {
+                let empty_diagnostics = create_empty_diagnostics(url.clone(), document);
+                diagnostics.insert(url.clone(), empty_diagnostics)
+            };
         }
 
         diagnostics
@@ -829,14 +698,17 @@ impl PestLanguageServerImpl {
             });
 
         self.documents = not_in_dir;
-        in_dir.iter().map(create_empty_diagnostics).collect()
+        in_dir
+            .into_iter()
+            .map(|(url, doc)| (url.clone(), create_empty_diagnostics(url, &doc)))
+            .collect()
     }
 
     async fn send_diagnostics(&self, diagnostics: Diagnostics) {
         for PublishDiagnosticsParams {
             uri,
             diagnostics,
-            version,
+            version
         } in diagnostics.into_values()
         {
             self.client
@@ -849,6 +721,5 @@ impl PestLanguageServerImpl {
         self.analyses
             .get(uri)
             .and_then(|analysis| analysis.rules.get(rule_name))
-            .and_then(|ra| ra.as_ref())
     }
 }
